requirements.txt
Package            Version
------------------ -----------
blinker            1.7.0
cachelib           0.10.2
certifi            2023.7.22
charset-normalizer 3.3.2
click              8.1.7
colorama           0.4.6
dnspython          2.4.2
email-validator    2.1.0.post1
Flask              3.0.0
Flask-Session      0.5.0
idna               3.4
itsdangerous       2.1.2
Jinja2             3.1.2
MarkupSafe         2.1.3
pip                23.3.1
python-dotenv      1.0.0
requests           2.31.0
urllib3            2.1.0
Werkzeug           3.0.1
